moved added logging to logging_utils
