+++ lm_eval/__main__.py
@@ -11,9 +11,9 @@ from typing import Union
 import numpy as np
 
 from lm_eval import evaluator, utils
-from lm_eval.logging_utils import WandbLogger
+from lm_eval.logging_utils import WandbLogger, add_env_info
 from lm_eval.tasks import TaskManager, include_path, initialize_tasks
-from lm_eval.utils import add_env_info, make_table
+from lm_eval.utils import make_table
 
 
 def _handle_non_serializable(o):
