+++ lm_eval/utils.py
@@ -10,12 +10,12 @@ import subprocess
 import sys
 from itertools import islice
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Optional
+from typing import Any, Callable, List
 
 import yaml
 from jinja2 import BaseLoader, Environment, StrictUndefined
-from torch.utils.collect_env import get_pretty_env_info
-from transformers import __version__ as trans_version
+
+from lm_eval.logging_utils import get_commit_from_path
 
 
 logging.basicConfig(
@@ -329,21 +329,6 @@ def run_task_tests(task_list: List[str]):
         )
 
 
-def get_commit_from_path(repo_path: Path) -> Optional[str]:
-    git_folder = Path(repo_path, ".git")
-    if git_folder.is_file():
-        git_folder = Path(
-            git_folder.parent, git_folder.read_text().split("\n")[0].split(" ")[-1]
-        )
-    if Path(git_folder, "HEAD").exists():
-        head_name = Path(git_folder, "HEAD").read_text().split("\n")[0].split(" ")[-1]
-        head_ref = Path(git_folder, head_name)
-        git_hash = head_ref.read_text().replace("\n", "")
-    else:
-        git_hash = None
-    return git_hash
-
-
 def get_git_commit_hash():
     """
     Gets the git commit hash of your current repo (if it exists).
@@ -356,23 +341,6 @@ def get_git_commit_hash():
         # FileNotFoundError occurs when git not installed on system
         git_hash = get_commit_from_path(os.getcwd())  # git hash of repo if exists
     return git_hash
-
-
-def add_env_info(storage: Dict[str, Any]):
-    try:
-        pretty_env_info = get_pretty_env_info()
-    except Exception as err:
-        pretty_env_info = str(err)
-    transformers_version = "Transformers: %s" % trans_version
-    upper_dir_commit = get_commit_from_path(
-        Path(os.getcwd(), "..")
-    )  # git hash of upper repo if exists
-    added_info = {
-        "pretty_env_info": pretty_env_info,
-        "transformers_version": transformers_version,
-        "upper_git_hash": upper_dir_commit,  # in case this repo is submodule
-    }
-    storage.update(added_info)
 
 
 def ignore_constructor(loader, node):
